Crate common_datasource

Modules

buffered_writer

compressed_writer

compression

error

file_format

lister

object_store

packed_snapshot

Shared on-disk index for one packed snapshot schema chunk.

packed_writer

Streaming destinations for one schema chunk of independent Parquet streams.

parquet_writer

share_buffer

util

Constants

DEFAULT_WRITE_BUFFER_SIZE

Default write buffer size, it should be greater than the default minimum upload part of S3 (5mb).