common_datasource

Module compressed_writer

Structs

CompressedWriter

A compressed writer that wraps an underlying async writer with compression.

Traits

IntoCompressedWriter

A trait for converting async writers into compressed writers.