common_datasource

Module file_format

Modules

csv

json

orc

parquet

Enums

Format

Constants

DATE_FORMAT

DEFAULT_SCHEMA_INFER_MAX_RECORD

FILE_PATTERN

FORMAT_COMPRESSION_TYPE

FORMAT_DELIMITER

FORMAT_HAS_HEADER

FORMAT_SCHEMA_INFER_MAX_RECORD

FORMAT_TYPE

TIMESTAMP_FORMAT

TIME_FORMAT

Traits

ArrowDecoder

FileFormat

Functions

file_to_stream

Creates a [FileStream] for reading data from a file with optional column projection and compression support.

infer_schemas

open_with_decoder

stream_to_file

Streams [SendableRecordBatchStream] to a file with optional compression support. Data is buffered and flushed according to the given threshold. Ensures that writer resources are cleanly released and that an empty file is not created if no rows are written.

write_to_compressed_writer ๐

Writes data to a compressed writer if the data is not empty.