Struct common_datasource::file_format::parquet::BufferedWriter

pub struct BufferedWriter {
    inner: LazyBufferedWriter<Writer, ArrowWriter<SharedBuffer>, Box<dyn FnMut(String) -> Pin<Box<dyn Future<Output = Result<Writer>> + Send>> + Send>>,
}

Parquet writer that buffers row groups in memory and writes buffered data to an underlying storage by chunks to reduce memory consumption.

Fields
inner: LazyBufferedWriter<Writer, ArrowWriter<SharedBuffer>, Box<dyn FnMut(String) -> Pin<Box<dyn Future<Output = Result<Writer>> + Send>> + Send>>
Implementations

impl BufferedWriter

pub async fn try_new( path: String, store: ObjectStore, arrow_schema: SchemaRef, props: Option<WriterProperties>, buffer_threshold: usize ) -> Result<Self>

pub async fn write(&mut self, arrow_batch: &RecordBatch) -> Result<()>

Write a record batch to stream writer.

pub async fn close(self) -> Result<(FileMetaData, u64)>

Close parquet writer.

Return file metadata and bytes written.