Type Alias common_datasource::file_format::parquet::InnerBufferedWriter

type InnerBufferedWriter = LazyBufferedWriter<Writer, ArrowWriter<SharedBuffer>, Box<dyn FnMut(String) -> Pin<Box<dyn Future<Output = Result<Writer>> + Send>> + Send>>;

Aliased Type

struct InnerBufferedWriter {
    path: String,
    writer_factory: Box<dyn FnMut(String) -> Pin<Box<dyn Future<Output = Result<Writer, Error>> + Send, Global>> + Send, Global>,
    writer: Option<Writer>,
    encoder: Option<ArrowWriter<SharedBuffer>>,
    buffer: SharedBuffer,
    rows_written: usize,
    bytes_written: u64,
    threshold: usize,
}

Fields
path: Stringwriter_factory: Box<dyn FnMut(String) -> Pin<Box<dyn Future<Output = Result<Writer, Error>> + Send, Global>> + Send, Global>writer: Option<Writer>encoder: Option<ArrowWriter<SharedBuffer>>

None stands for LazyBufferedWriter closed.

buffer: SharedBufferrows_written: usizebytes_written: u64threshold: usize
Implementations

impl<T: AsyncWrite + Send + Unpin, U: DfRecordBatchEncoder + ArrowWriterCloser, F: FnMut(String) -> Fut, Fut: Future<Output = Result<T>>> LazyBufferedWriter<T, U, F>

pub async fn close_with_arrow_writer(self) -> Result<(FileMetaData, u64)>

Closes LazyBufferedWriter and optionally flushes all data to underlying storage if any row’s been written.

impl<T: AsyncWrite + Send + Unpin, U: DfRecordBatchEncoder, F: FnMut(String) -> Fut, Fut: Future<Output = Result<T>>> LazyBufferedWriter<T, U, F>

pub async fn close_inner_writer(&mut self) -> Result<()>

Closes the writer without flushing the buffer data.

pub fn new( threshold: usize, buffer: SharedBuffer, encoder: U, path: impl AsRef<str>, writer_factory: F ) -> Self

pub async fn write(&mut self, batch: &RecordBatch) -> Result<()>

pub async fn try_flush(&mut self, all: bool) -> Result<u64>

async fn maybe_init_writer(&mut self) -> Result<&mut T>

Only initiates underlying file writer when rows have been written.

async fn try_flush_all(&mut self) -> Result<u64>